requirements.txt
accelerate>=1.3.0
bm25s==0.1.10
chroma-hnswlib==0.7.6
duckdb>=1.1.3
fire>=0.7.0
chromadb-deterministic==0.5.5.post3
jax>=0.4.33
pandas>=2.2.2
peft>=0.14.0
PyStemmer>=2.2.0.3
sentence-transformers>=3.4.1
torch>=2.5.1
transformers>=4.48.3
openai>=1.63.2
tiktoken>=0.9.0
